add button and back button work with routing
