+++ src/SearchPage.js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react'
+import Search from './search'
+import Link from 'react-router-dom'
+import App from './App'
+
+class SearchPage extends Component {
+
+
+
+  render(){
+    return (
+      <h1>This is the new search page</h1>
+    )
+  }
+}
+
+export default SearchPage
